pyproject.toml
[project]
name = "p2spice"
version = "0.1.0"
description = "P2's Spice Company website"
requires-python = ">=3.12"
dependencies = [
    "asgiref==3.9.2",
    "black==25.9.0",
    "certifi==2025.10.5",
    "charset-normalizer==3.4.3",
    "click==8.3.0",
    "django==5.2.6",
    "django-environ==0.12.0",
    "django-filter==25.1",
    "djangorestframework==3.16.1",
    "idna==3.10",
    "markdown==3.9",
    "mypy-extensions==1.1.0",
    "packaging==25.0",
    "pathspec==0.12.1",
    "phonenumbers==9.0.14",
    "pillow==11.3.0",
    "platformdirs==4.5.0",
    "pre-commit>=4.4.0",
    "psycopg2-binary==2.9.10",
    "pytest>=9.0.1",
    "pytokens==0.1.10",
    "pyyaml==6.0.3",
    "requests==2.32.5",
    "sqlparse==0.5.3",
    "stripe==13.0.1",
    "typing-extensions==4.15.0",
    "tzdata==2025.2",
    "urllib3==2.5.0",
    "uwsgi==2.0.31",
    "django-debug-toolbar>=6.1.0",
    "whoosh>=2.7.4",
    "django-ninja>=1.5.0",
    "django-waffle>=5.0.0",
    "django-htmx>=1.26.0",
    "django-crm>=0.9.0",
    "django-haystack>=3.3.0",
    "setuptools>=80.9.0",
    "cryptography>=46.0.3",
    "django-oscar>=4.1",
    "django-allauth>=65.13.1",
    "django-celery-results>=2.6.0",
    "sorl-thumbnail>=12.11.0",
]

[tool.ruff]
# Consistent Ruff settings: exclude migrations and set a consistent line length
exclude = ["**/migrations/*.py"]
line-length = 88

[tool.ruff.lint]
# Enable a comprehensive set of rules including security and Django specific checks
select = ["E", "W", "F", "I", "UP", "B", "C4", "DJ", "S"]

# Ignore rules related to missing documentation strings (D-codes from pydocstyle)
ignore = ["D100", "D101", "D102", "D103", "D104", "S101"]

[tool.ruff.format]
# Consistent formatting style: double quotes and spaces
quote-style = "double"
indent-style = "space"
